Added report system

+++ l4/Parallel/source/MyBubbleSort.c
@@ -8,14 +8,14 @@
  * 
  */
 
-long long int MyBubbleSort(array* unsortedArray){
-	long long int counter=0;
-	long long int size=unsortedArray->length;
+int MyBubbleSort(array* unsortedArray){
+	int counter=0;
+	int size=unsortedArray->length;
 	double *arr=unsortedArray->Arr;
 	double buf=0.;
 	
-	for(long long int i=0;i<size;++i)
-		for(long long int j=size-1;j>i;--j){
+	for(int i=0;i<size;++i)
+		for(int j=size-1;j>i;--j){
 			if(arr[j]<arr[j-1]){
 				counter++;
 				buf=arr[j];
