+++ l4/Parallel/source/MyNormalizator.c
@@ -7,10 +7,10 @@
  * @return
  * 
  */
-long long int MyNormalizator(array* firstArray){
-	long long int sentinel=0; //индекс первого неотсортированного элемента
+int MyNormalizator(array* firstArray){
+	int sentinel=0; //индекс первого неотсортированного элемента
 	double buf=0.;
-	long long int counter=0; //число переcтановок
+	int counter=0; //число переcтановок
 
 	if(firstArray->Arr!=NULL){
 	for(;sentinel<(firstArray->length-1) && 
@@ -19,7 +19,7 @@ long long int MyNormalizator(array* firstArray){
 		; /*VOID*/
 
 	for(sentinel++;sentinel < firstArray->length;sentinel++){
-		for(long long int j=sentinel-1;j>=0;j--){
+		for(int j=sentinel-1;j>=0;j--){
 			if(firstArray->Arr[j+1]<firstArray->Arr[j]){
 				counter++;
 				buf=firstArray->Arr[j+1];
