+++ l4/Parallel/source/getNinjaIdx.c
@@ -8,9 +8,9 @@
  * @return индекс первого элемента большего чем ведущий
  * 
  */
-long long int getNinjaIdx(array* wholeArray, double lider){
+int getNinjaIdx(array* wholeArray, double lider){
 	double *temp=wholeArray->Arr;
-	long long int i=0;
+	int i=0;
 	
 	if(temp!=NULL)
 		for(;(*temp<lider)&&(i<(wholeArray->length));temp++, i++);
